Refactor azorious/snapshot proposal duality handling. Tighten up types, remove prop drilling
